+++ collective/tablepage/fields/file.py
@@ -14,12 +14,21 @@ from plone.memoize.instance import memoize
 from zExceptions import BadRequest
 from zope.component import getMultiAdapter
 from zope.interface import implements
-
+from zope.lifecycleevent import modified
 try:
     from zope.browserpage.viewpagetemplatefile import ViewPageTemplateFile
 except ImportError:
     # Plone < 4.1
     from zope.app.pagetemplate.viewpagetemplatefile import ViewPageTemplateFile
+
+try:
+    from plone.dexterity.interfaces import IDexterityContent
+    from plone.namedfile.file import NamedBlobFile
+    HAS_DEXTERITY = True
+except ImportError:
+    HAS_DEXTERITY = False
+    IDexterityContent = None
+    NamedBlobFile = None
 
 TYPE_TO_CREATE = 'File'
 
@@ -27,9 +36,24 @@ TYPE_TO_CREATE = 'File'
 class FileField(BaseField):
     implements(IFileColumnField)
 
+    portal_type = TYPE_TO_CREATE
+
     edit_template = ViewPageTemplateFile('templates/file.pt')
     view_template = ViewPageTemplateFile('templates/file_view.pt')
     cache_time = 60 * 60 # 2 hours
+
+    def _get_obj_info(self, uuid):
+        # for fields that need to refer to other contents
+        info = super(FileField, self)._get_obj_info(uuid)
+        if 'url' in info:
+            url = info['url'].rstrip('/')
+            if HAS_DEXTERITY and IDexterityContent.providedBy(info['object']):
+                download_url = url + '/download'
+            else:
+                download_url = url + '/at_download/file'
+            info['download_url'] = download_url
+
+        return info
 
     def render_view(self, data, index=None, storage=None):
         self.data = data or ''
@@ -41,7 +65,7 @@ class FileField(BaseField):
 
     def can_add_file(self):
         member = getMultiAdapter((self.context, self.request), name=u'plone_portal_state').member()
-        return member.has_permission(permissions[TYPE_TO_CREATE], self.attachment_storage)
+        return member.has_permission(permissions[self.portal_type], self.attachment_storage)
 
     @property
     @memoize
@@ -56,7 +80,7 @@ class FileField(BaseField):
 
     def attachments(self):
         catalog = getToolByName(self.context, 'portal_catalog')
-        files_in_storage = catalog(portal_type='File',
+        files_in_storage = catalog(portal_type=self.portal_type,
                                    path={'query': '/'.join(self.attachment_storage.getPhysicalPath()),
                                          'depth': 1,
                                          },
@@ -66,7 +90,7 @@ class FileField(BaseField):
             # when editing we haven't the old file still there
             old = catalog(UID=self.data)
             if old and old[0].getPath() not in [a.getPath() for a in files_in_storage]:
-                return old + files_in_storage 
+                return old + files_in_storage
         return files_in_storage
 
 
@@ -99,7 +123,7 @@ class MultipleFilesField(FileField):
         catalog = getToolByName(self.context, 'portal_catalog')
         files_in_storage = []
         for uid in filterUIDs:
-            file_in_storage = catalog(portal_type='File',
+            file_in_storage = catalog(portal_type=self.portal_type,
                                       path={'query': '/'.join(self.attachment_storage.getPhysicalPath()),
                                             'depth': 1,
                                             },
@@ -117,7 +141,10 @@ class FileDataRetriever(LinkedObjectFinder):
      * return the UID of the new File
     """
 
+    portal_type = TYPE_TO_CREATE
     implements(IColumnDataRetriever)
+    field_name = 'file'
+    field_value_class = NamedBlobFile
 
     def __init__(self, context):
         self.context = context
@@ -129,10 +156,10 @@ class FileDataRetriever(LinkedObjectFinder):
             title = request.get('title_%s' % name)
             description = request.get('description_%s' % name)
             file = request.get(name)
-            newId = folder.generateUniqueId(TYPE_TO_CREATE)
+            newId = folder.generateUniqueId(self.portal_type)
+            plone_utils = getToolByName(self.context, 'plone_utils')
             if not title and file.filename in folder.objectIds():
                 # WARNING: we don't get the file title, to obtain the id
-                plone_utils = getToolByName(self.context, 'plone_utils')
                 plone_utils.addPortalMessage(_('duplicate_file_error_with_link',
                                                default=u'There is already an item named ${name} in this folder.\n'
                                                        u'Loading of the new attachment has been aborted '
@@ -140,23 +167,31 @@ class FileDataRetriever(LinkedObjectFinder):
                                                mapping={'name': file.filename}),
                                              type='warning')
                 return {name: folder[file.filename].UID()}
-            folder.invokeFactory(id=newId, type_name=TYPE_TO_CREATE,
+            folder.invokeFactory(id=newId, type_name=self.portal_type,
                                  title=title, description=description)
             new_doc = folder[newId]
-            # force rename (processForm will not work with files)
-            new_doc._renameAfterCreation()
-            # this will trigger proper lifecycle events
-            new_doc.processForm()
-            try:
-                new_doc.edit(file=file)
-            except BadRequest:
-                # Still don't get how, but sometimes this happen (at least on Plone 3)
-                plone_utils.addPortalMessage(_('duplicate_file_critical_error',
-                                               default=u'There is already an item named ${name} in this folder.\n'
-                                                       u'Loading of the attachment has been aborted.',
-                                               mapping={'name': file.filename}),
-                                             type='error')
-                return None
+            if HAS_DEXTERITY and IDexterityContent.providedBy(new_doc):
+                setattr(new_doc, self.field_name, self.field_value_class(file, filename=unicode(file.filename)))
+                modified(new_doc)
+                if not folder.has_key(file.filename):
+                    folder.manage_renameObject(new_doc.getId(), file.filename)
+            else:
+                # force rename (processForm will not work with files)
+                new_doc._renameAfterCreation()
+                # this will trigger proper lifecycle events
+                new_doc.processForm()
+
+                try:
+                    new_doc.edit(**{self.field_name: file})
+
+                except BadRequest:
+                    # Still don't get how, but sometimes this happen (at least on Plone 3)
+                    plone_utils.addPortalMessage(_('duplicate_file_critical_error',
+                                                   default=u'There is already an item named ${name} in this folder.\n'
+                                                           u'Loading of the attachment has been aborted.',
+                                                   mapping={'name': file.filename}),
+                                                 type='error')
+                    return None
             return {name: new_doc.UID()}
         elif request.get("existing_%s" % name):
             return {name: request.get("existing_%s" % name)}
@@ -193,6 +228,10 @@ class MultipleFilesDataRetriever(LinkedObjectFinder):
      * return a list of UID
     """
 
+    portal_type = TYPE_TO_CREATE
+    field_name = 'file'
+    field_value_class = NamedBlobFile
+
     implements(IColumnDataRetriever)
 
     def __init__(self, context):
@@ -214,7 +253,7 @@ class MultipleFilesDataRetriever(LinkedObjectFinder):
                 description = request.get('description_%s_%s' % (name, cnt))
                 file = request.get("%s_%s" % (name, cnt))
                 cnt += 1
-                newId = folder.generateUniqueId(TYPE_TO_CREATE)
+                newId = folder.generateUniqueId(self.portal_type)
                 if not title and file.filename in folder.objectIds():
                     # WARNING: we don't get the file title, to obtain the id
                     plone_utils.addPortalMessage(_('duplicate_file_error',
@@ -224,14 +263,20 @@ class MultipleFilesDataRetriever(LinkedObjectFinder):
                                                  type='warning')
                     results.append(folder[file.filename].UID())
                     continue
-                folder.invokeFactory(id=newId, type_name=TYPE_TO_CREATE,
+                folder.invokeFactory(id=newId, type_name=self.portal_type,
                                      title=title, description=description)
                 new_doc = folder[newId]
-                # force rename (processForm will not work with files)
-                new_doc._renameAfterCreation()
-                # this will trigger proper lifecycle events
-                new_doc.processForm()
-                new_doc.edit(file=file)
+                if HAS_DEXTERITY and IDexterityContent.providedBy(new_doc):
+                    setattr(new_doc, self.field_name, self.field_value_class(file, filename=unicode(file.filename)))
+                    modified(new_doc)
+                    if not folder.has_key(file.filename):
+                        folder.manage_renameObject(new_doc.getId(), file.filename)
+                else:
+                    # force rename (processForm will not work with files)
+                    new_doc._renameAfterCreation()
+                    # this will trigger proper lifecycle events
+                    new_doc.processForm()
+                new_doc.edit(**{self.field_name: file})
                 results.append(new_doc.UID())
             else:
                 break