+++ docs/HISTORY.rst
@@ -4,6 +4,8 @@ Changelog
 0.10.7 (unreleased)
 -------------------
 
+- FileField is Dexterity compatible [thomasdesvenain]
+- Added ImageField [thomasdesvenain]
 - We can set a default ordering [thomasdesvenain]
 - Number field HTML5 validator now allows negative numbers (fixes #29) [petri]